pyproject.toml
[project]
name = "weather-mcp-server"
version = "0.1.0"
description = "A weather MCP server built with FastMCP"
authors = [
    { name = "qianyan", email = "qianyan.lambda@gmail.com" }
]
dependencies = [
    "fastmcp>=0.4.1",
    "requests==2.32.3",
    "python-dotenv==1.1.0",
    "pydantic==2.11.1",
]
requires-python = ">=3.13"
readme = "README.md"
license = { text = "MIT" }
keywords = ["weather", "mcp", "claude", "fastmcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/qianyan/weather-mcp-server"
Repository = "https://github.com/qianyan/weather-mcp-server.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "ruff>=0.11.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/weather_mcp_server"]

[tool.ruff]
line-length = 88
target-version = "py38"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[project.scripts]
weather-mcp-server = "weather_mcp_server.server:main" 